Added Exit Polling
Fairly basic callback added once the window is closed. Now that I understand the general pattern of GLFW window callbacks, it'll be easy to add more functionality to them as needed in the future. To be honest, I thought it'd be more complicated, since SFML's event polling system is slightly more involved than this one seems to be.