Cleaned up Allocate's invocation
I made it really easy and pretty to use Allocate now! Well, at least, for the client. It was a PITA to get things where they needed to be, and I really don't like exposing code in my header. I'll keep the client-side Allocate function there, though, since it's templated (and templated stuff is just better kept in the header). It looks even worse with the deallocate function whose definition can't be shown (otherwise it'll cause redefinition-related linker errors).